Refactor unrelated code into separate util
